Conflicts:
	README.md
	src/build/verinfo.py
	src/hbmame/osd/windows/hbmame.man
	src/mame/osd/windows/mame/mame.ico
	src/mame/osd/windows/mess/mess.ico
	src/mame/osd/windows/mess/mess.man
	src/mame/osd/windows/mess/mess.rc
	src/mess/osd/windows/mess.man

+++ src/osd/modules/sound/sdl_sound.c
@@ -1,11 +1,8 @@
-// license:???
-// copyright-holders:???
+// license:BSD-3-Clause
+// copyright-holders:Olivier Galibert, R. Belmont
 //============================================================
 //
 //  sound.c - SDL implementation of MAME sound routines
-//
-//  Copyright (c) 1996-2010, Nicola Salmoria and the MAME Team.
-//  Visit http://mamedev.org for licensing and usage restrictions.
 //
 //  SDLMAME by Olivier Galibert and R. Belmont
 //
@@ -187,7 +184,7 @@ void sound_sdl::unlock_buffer(void)
 
 void sound_sdl::att_memcpy(void *dest, const INT16 *data, int bytes_to_copy)
 {
-	int level= (int) (pow(10.0, (float) attenuation / 20.0) * 128.0);
+	int level= (int) (pow(10.0, (double) attenuation / 20.0) * 128.0);
 	INT16 *d = (INT16 *) dest;
 	int count = bytes_to_copy/2;
 	while (count>0)
